mbedtls 3.6

+++ lib/luatos-soc-2022/interface/src/luat_camera_ec618.c
@@ -29,84 +29,84 @@ extern void CSPI_Stop(uint8_t I2SID);
 
 typedef struct
 {
-	CBFuncEx_t callback;
-	void *param;
-	uint32_t total_byte;
-	uint8_t is_init;
-	uint8_t is_running;
+    CBFuncEx_t callback;
+    void *param;
+    uint32_t total_byte;
+    uint8_t is_init;
+    uint8_t is_running;
 }luat_camera_ctrl_t;
 
 static luat_camera_ctrl_t g_s_camera[I2S_MAX];
 
-static int luat_camera_dummy_callback(void *pdata, void *param)
+static long luat_camera_dummy_callback(void *pdata, void *param)
 {
-	return 0;
+    return 0;
 }
 
 
 
 int luat_camera_setup(int id, luat_spi_camera_t *conf, void * callback, void *param)
 {
-	if (id < 0 || id >= I2S_MAX || !conf || !callback) return -ERROR_PARAM_INVALID;
-	if (g_s_camera[id].is_init) return -ERROR_OPERATION_FAILED;
-	uint16_t color_byte = conf->only_y?1:2;
-	uint16_t one_buf_height = 8000 / (conf->sensor_width * color_byte);
-//	DBG("one buf height %d", conf->one_buf_height);
-	g_s_camera[id].total_byte = one_buf_height * conf->sensor_width * color_byte;
-	g_s_camera[id].callback = callback;
-	g_s_camera[id].param = param;
+    if (id < 0 || id >= I2S_MAX || !conf || !callback) return -ERROR_PARAM_INVALID;
+    if (g_s_camera[id].is_init) return -ERROR_OPERATION_FAILED;
+    uint16_t color_byte = conf->only_y?1:2;
+    uint16_t one_buf_height = 8000 / (conf->sensor_width * color_byte);
+//  DBG("one buf height %d", conf->one_buf_height);
+    g_s_camera[id].total_byte = one_buf_height * conf->sensor_width * color_byte;
+    g_s_camera[id].callback = callback;
+    g_s_camera[id].param = param;
 
-	if (id)
-	{
-		GPIO_IomuxEC618(18, 1, 1, 0);
-		GPIO_IomuxEC618(19, 1, 1, 0);
-		GPIO_IomuxEC618(21, 1, 1, 0);
-		GPIO_IomuxEC618(22, 1, 1, 0);
-	}
-	else
-	{
-		GPIO_IomuxEC618(39, 1, 1, 0);
-		GPIO_IomuxEC618(35, 1, 1, 0);
-		GPIO_IomuxEC618(37, 1, 1, 0);
-		GPIO_IomuxEC618(38, 1, 1, 0);
-	}
+    if (id)
+    {
+        GPIO_IomuxEC618(18, 1, 1, 0);
+        GPIO_IomuxEC618(19, 1, 1, 0);
+        GPIO_IomuxEC618(21, 1, 1, 0);
+        GPIO_IomuxEC618(22, 1, 1, 0);
+    }
+    else
+    {
+        GPIO_IomuxEC618(39, 1, 1, 0);
+        GPIO_IomuxEC618(35, 1, 1, 0);
+        GPIO_IomuxEC618(37, 1, 1, 0);
+        GPIO_IomuxEC618(38, 1, 1, 0);
+    }
 
-	CSPI_Setup(id, conf->camera_speed, conf->spi_mode, conf->is_msb, conf->is_two_line_rx, conf->only_y, conf->seq_type, conf->rowScaleRatio, conf->colScaleRatio, conf->scaleBytes);
+    CSPI_Setup(id, conf->camera_speed, conf->spi_mode, conf->is_msb, conf->is_two_line_rx, conf->only_y, conf->seq_type, conf->rowScaleRatio, conf->colScaleRatio, conf->scaleBytes);
 
-	g_s_camera[id].is_running = 0;
-	g_s_camera[id].is_init = 1;
+    g_s_camera[id].is_running = 0;
+    g_s_camera[id].is_init = 1;
 
-	return 0;
+    return 0;
 }
 
 
 int luat_camera_start(int id)
 {
-	if (id < 0 || id >= I2S_MAX) return -1;
-	if (!g_s_camera[id].is_init || g_s_camera[id].is_running) return -ERROR_OPERATION_FAILED;
-	CSPI_Rx(id, g_s_camera[id].total_byte, g_s_camera[id].callback, g_s_camera[id].param);
-	g_s_camera[id].is_running = 1;
-	return 0;
+    if (id < 0 || id >= I2S_MAX) return -1;
+    if (!g_s_camera[id].is_init || g_s_camera[id].is_running) return -ERROR_OPERATION_FAILED;
+    CSPI_Rx(id, g_s_camera[id].total_byte, g_s_camera[id].callback, g_s_camera[id].param);
+    g_s_camera[id].is_running = 1;
+    return 0;
 }
 
 
 int luat_camera_stop(int id)
 {
-	if (id < 0 || id >= I2S_MAX) return -1;
-	if (!g_s_camera[id].is_init || !g_s_camera[id].is_running) return -ERROR_OPERATION_FAILED;
-	CSPI_RxStop(id);
-	g_s_camera[id].is_running = 0;
-	return 0;
+    if (id < 0 || id >= I2S_MAX) return -1;
+    if (!g_s_camera[id].is_init || !g_s_camera[id].is_running) return -ERROR_OPERATION_FAILED;
+    CSPI_RxStop(id);
+    g_s_camera[id].is_running = 0;
+    return 0;
 }
 
 
 int luat_camera_close(int id)
 {
-	if (id < 0 || id >= I2S_MAX) return -1;
-	if (!g_s_camera[id].is_init) return -ERROR_OPERATION_FAILED;
-	CSPI_Stop(id);
-	g_s_camera[id].callback = luat_camera_dummy_callback;
-	return 0;
+    if (id < 0 || id >= I2S_MAX) return -1;
+    if (!g_s_camera[id].is_init) return -ERROR_OPERATION_FAILED;
+    CSPI_Stop(id);
+    g_s_camera[id].callback = luat_camera_dummy_callback;
+    return 0;
 
 }
